requirements.txt
boto3== 1.37.18 #
Flask == 3.1.0 #
flask-cors == 5.0.1 #
Flask-SQLAlchemy == 3.1.1
mysql-connector-python == 9.2.0 #
SQLAlchemy == 2.0.39 #
gunicorn  == 23.0.0
PyJWT == 2.10.1
bcrypt == 4.3.0